requirements.txt
# ============================================
# Coffeehouse Face Recognition System
# Python Dependencies
# ============================================

# Face Recognition & Image Processing
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

# Database
pymongo==4.6.0

# HTTP API Server (for Mobile App)
flask==3.0.0
flask-cors==4.0.0

# Configuration
python-dotenv==1.0.0

# ============================================
# System Dependencies (Install BEFORE pip install)
# ============================================
#
# Ubuntu/Debian:
#   sudo apt-get update
#   sudo apt-get install -y \
#       build-essential \
#       cmake \
#       libopenblas-dev \
#       liblapack-dev \
#       libx11-dev \
#       libgtk-3-dev \
#       pkg-config
#
# macOS:
#   brew install cmake dlib
#
# ============================================
# Installation Steps
# ============================================
#
# 1. Install system dependencies (see above)
# 2. Create virtual environment: python3 -m venv venv
# 3. Activate virtual environment: source venv/bin/activate
# 4. Upgrade pip: pip install --upgrade pip
# 5. Install dependencies: pip install -r requirements.txt
#
# ============================================
# Troubleshooting
# ============================================
#
# If face-recognition installation fails:
#   - Ensure CMake is installed: cmake --version
#   - Install system dependencies first
#   - Try: pip install --no-cache-dir face-recognition==1.3.0
#
# If dlib build fails:
#   - Check system dependencies are installed
#   - On Ubuntu: sudo apt-get install python3-dev
#   - On macOS: xcode-select --install
#
